requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# CLIP
ftfy
regex
tqdm

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization (optional)
matplotlib>=3.7.0

# Development tools (optional)
jupyter
ipython
torchsummary

# Additional utilities
timm